[TensorPipe] Use registry for transports and channels
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39957

In order to provide a pluggable and extendable way to add new transports and channels to the TensorPipe agent, we use two registries. This allows us to separate the specific details of each backend (e.g., how it determines what address to use) from the generic logic of setting up TensorPipe.

Differential Revision: [D22017614](https://our.internmc.facebook.com/intern/diff/D22017614/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D22017614/)!
ghstack-source-id: 105875475